refactor(all): split into primitives, decode, encode files
